added password request before deleting recipe

+++ src/components/Notification/Notification.js
@@ -1,8 +1,8 @@
-import { useContext } from 'react';
+import { useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { doc, deleteDoc } from 'firebase/firestore';
 
-import { db } from '../../services/Firebase/firebase';
+import { db, getPass } from '../../services/Firebase/firebase';
 
 import { NotificationContext } from '../../context/NotificationContext';
 
@@ -11,13 +11,22 @@ import './notification.css';
 const Notification = () => {
   const { notification, setNotification, style, setMessage, recipeId } =
     useContext(NotificationContext);
+  const [dataPass, setDataPass] = useState('');
+  const [pass, setPass] = useState('');
   const navigate = useNavigate();
 
   if (notification.message === '') {
     return null;
   }
 
+  getPass().then((data) => setDataPass(data));
+
   const deleteRecipe = async () => {
+    if (pass !== dataPass || pass === '') {
+      console.log('incorrect pass');
+      return;
+    }
+
     await deleteDoc(doc(db, 'recipes', recipeId));
 
     navigate('/');
@@ -30,6 +39,12 @@ const Notification = () => {
       <div className="notif-overlay">
         <div className="confirm-notif">
           <h4>{notification.message}</h4>
+
+          <input
+            className="notif-input"
+            type="text"
+            onChange={(e) => setPass(e.target.value)}
+          />
 
           <div className="notif-btn-container">
             <button className="notif-btn" onClick={() => setMessage('')}>
